+++ 3denEnhanced/attributesObject/SPRTickets.hpp
@@ -5,7 +5,6 @@ class ENH_SPR_Tickets: ENH_SetCaptive
 	property = "ENH_SPR_Tickets";
 	control = "EditShort"
 	expression = "if (!is3DEN && {!isMultiplayer}) then {_this setVariable ['%s',_value]}";
-	condition = "objectBrain";
 	typeName = "NUMBER";
 	defaultValue = "0";
 };